+++ modules/portal/portal-sso-token/src/com/liferay/portal/sso/token/internal/security/auth/CookieTokenRetriever.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.sso.token.internal.auth;
+package com.liferay.portal.sso.token.internal.security.auth;
 
-import com.liferay.portal.sso.token.auth.TokenLocation;
-import com.liferay.portal.sso.token.auth.TokenRetriever;
+import com.liferay.portal.kernel.util.CookieKeys;
+import com.liferay.portal.sso.token.security.auth.TokenLocation;
+import com.liferay.portal.sso.token.security.auth.TokenRetriever;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -25,18 +26,18 @@ import org.osgi.service.component.annotations.Component;
  * @author Michael C. Han
  */
 @Component(immediate = true, service = TokenRetriever.class)
-public class RequestHeaderTokenRetriever implements TokenRetriever {
+public class CookieTokenRetriever implements TokenRetriever {
 
 	@Override
 	public String getLoginToken(
 		HttpServletRequest request, String userTokenName) {
 
-		return request.getHeader(userTokenName);
+		return CookieKeys.getCookie(request, userTokenName, false);
 	}
 
 	@Override
 	public TokenLocation getTokenLocation() {
-		return TokenLocation.REQUEST_HEADER;
+		return TokenLocation.COOKIE;
 	}
 
 }
